+++ .eslintrc.js
@@ -2,7 +2,7 @@ module.exports = {
   plugins: ['node'],
   extends: ['eslint:recommended', 'plugin:node/recommended'],
   parserOptions: {
-    ecmaVersion: 2017,
+    ecmaVersion: 2018,
     sourceType: 'script',
     ecmaFeatures: {
       experimentalObjectRestSpread: true
@@ -16,6 +16,8 @@ module.exports = {
   rules: {
     'no-console': 0,
     semi: [2, 'never'],
-    quotes: [2, 'single', 'avoid-escape']
+    quotes: [2, 'single', 'avoid-escape'],
+    complexity: 0,
+    duplication: 0
   }
 }