+++ src/api/snapshot-metrics.js
@@ -1,5 +1,4 @@
-const gql = require('../utils/api-client')
-const { handleError } = require('../utils/api-error')
+const { request } = require('./graphql')
 
 const SNAPSHOT_METRICS_QUERY = `
   query GetSnapshotMetrics(
@@ -52,7 +51,80 @@ const SNAPSHOT_METRICS_QUERY = `
   }
 `
 
+const TIMESERIES_FRAGMENT = `
+timeseries(duration_in_days: $durationInDays, metrics: $metrics) {
+  snapshots {
+    id
+    sequenceId: iid
+    createdAt
+  }
+
+  series {
+    metric {
+      name
+      label
+      formatter
+      docsPath
+    }
+
+    sets {
+      page {
+        uuid
+        name
+        url
+      }
+
+      profile {
+        uuid
+        name
+      }
+
+      values {
+        snapshot
+        value
+      }
+    }
+  }
+}`
+
 const PULSE_METRICS_QUERY = `
+  query GetPulseData(
+    $site: String!
+    $durationInDays: Int
+    $metrics: [MetricTag!]
+  ) {
+    organisation {
+      site(slug: $site) {
+        hasRecentlyCompletedSnapshots
+
+        ${TIMESERIES_FRAGMENT}
+
+        pages {
+          name
+          uuid
+          url
+        }
+
+        testProfiles {
+          name
+          uuid
+          device {
+            title
+          }
+          bandwidth {
+            title
+          }
+          isMobile
+          jsIsDisabled
+          hasDeviceEmulation
+          hasBandwidthEmulation
+        }
+      }
+    }
+  }
+`
+
+const PULSE_PAGE_METRICS_QUERY = `
   query GetPulsePageData(
     $site: String!
     $page: String
@@ -61,45 +133,14 @@ const PULSE_METRICS_QUERY = `
   ) {
     organisation {
       site(slug: $site) {
-        hasCompletedSnapshots
+        hasRecentlyCompletedSnapshots
 
         page(uuid: $page) {
-          name
-          uuid
-          url
-
-          timeseries(duration_in_days: $durationInDays, metrics: $metrics) {
-            snapshots {
-              id
-              sequenceId: iid
-              createdAt
-            }
-
-            series {
-              metric {
-                name
-                label
-                formatter
-                docsPath
-              }
-
-              sets {
-                profile {
-                  id
-                  name
-                }
-
-                values {
-                  snapshot
-                  value
-                }
-              }
-            }
-          }
+          ${TIMESERIES_FRAGMENT}
         }
 
         testProfiles {
-          id
+          uuid
           name
           device {
             title
@@ -118,29 +159,33 @@ const PULSE_METRICS_QUERY = `
 `
 
 const snapshot = async ({ site, snapshotId }) => {
-  try {
-    const response = await gql.request(SNAPSHOT_METRICS_QUERY, {
-      site,
-      snapshotId
-    })
-    return response.organisation.site
-  } catch (e) {
-    return handleError(e)
-  }
+  const response = await request({
+    query: SNAPSHOT_METRICS_QUERY,
+    site,
+    snapshotId
+  })
+  return response.organisation.site
 }
 
 const pulse = async ({ site, page, durationInDays, metrics }) => {
-  try {
-    const response = await gql.request(PULSE_METRICS_QUERY, {
+  let query,
+    attributes = {}
+
+  if (page) {
+    query = PULSE_PAGE_METRICS_QUERY
+    attributes = {
       site,
       page,
       durationInDays,
       metrics
-    })
-    return response.organisation.site
-  } catch (e) {
-    return handleError(e)
+    }
+  } else {
+    query = PULSE_METRICS_QUERY
+    attributes = { site, durationInDays, metrics }
   }
+
+  const response = await request({ query: query, ...attributes })
+  return response.organisation.site
 }
 
 module.exports = {