+++ src/api/snapshot.js
@@ -1,5 +1,4 @@
-const gql = require('../utils/api-client')
-const { handleError } = require('../utils/api-error')
+const { request } = require('./graphql')
 
 const CREATE_MUTATION = `
   mutation CreateSnapshot(
@@ -16,8 +15,8 @@ const CREATE_MUTATION = `
 `
 
 const DELETE_MUTATION = `
-  mutation DeleteSnapshot($site:String!, $iid: String!){
-    deleteSnapshot(site: $site, iid: $iid) {
+  mutation DeleteSnapshot($site:String!, $id: String!){
+    deleteSnapshot(site: $site, iid: $id) {
       iid
     }
   }
@@ -42,41 +41,82 @@ const LIST_QUERY = `
   }
 `
 
-const create = async ({ site, ref }) => {
-  try {
-    const response = await gql.request(CREATE_MUTATION, {
-      site,
-      ref: new String(ref)
-    })
-    return response.createSnapshot
-  } catch (e) {
-    return handleError(e)
+const GET_SNAPSHOT_ARTIFACT_URLS = `
+  query GetSnapshotArtifacts($site: String!, $id: Int!) {
+    organisation {
+      site(slug: $site) {
+        pages {
+          uuid
+          name
+        }
+
+        testProfiles {
+          uuid
+          name
+        }
+
+        snapshot(id: $id) {
+          iid
+
+          tests {
+            status
+
+            har: artifactURL(name: TEST_ARTIFACT_HAR)
+            lighthouse: artifactURL(name: TEST_ARTIFACT_LIGHTHOUSE)
+            image: mediaURL(name: TEST_MEDIA_IMAGE)
+            gif: mediaURL(name:TEST_MEDIA_GIF)
+            video: mediaURL(name: TEST_MEDIA_VIDEO)
+
+            page {
+              uuid
+            }
+
+            testProfile {
+              uuid
+            }
+          }
+        }
+      }
+    }
   }
+`
+
+const create = async ({ site, ref }) => {
+  const response = await request({
+    query: CREATE_MUTATION,
+    site,
+    ref: new String(ref)
+  })
+  return response.createSnapshot
 }
 
-const destroy = async ({ site, iid }) => {
-  try {
-    const response = await gql.request(DELETE_MUTATION, {
-      site,
-      iid: String(iid)
-    })
-    return response.deleteSnapshot
-  } catch (e) {
-    return handleError(e)
-  }
+const destroy = async ({ site, id }) => {
+  const response = await request({
+    query: DELETE_MUTATION,
+    site,
+    id: Number(id)
+  })
+  return response.deleteSnapshot
 }
 
 const list = async ({ site }) => {
-  try {
-    const response = await gql.request(LIST_QUERY, { site })
-    return response.organisation.site.snapshots
-  } catch (e) {
-    return handleError(e)
-  }
+  const response = await request({ query: LIST_QUERY, site })
+  return response.organisation.site.snapshots
+}
+
+const fetchArtifacts = async ({ site, id }) => {
+  const response = await request({
+    query: GET_SNAPSHOT_ARTIFACT_URLS,
+    site,
+    id: Number(id)
+  })
+
+  return response.organisation.site
 }
 
 module.exports = {
   create,
   destroy,
-  list
+  list,
+  fetchArtifacts
 }