+++ src/cli.js
@@ -22,24 +22,26 @@ module.exports = require('yargs')
     'Options:': chalk.grey('Options:\n')
   })
   .version(pkg.version)
-  .showHelpOnFail(true)
   .recommendCommands()
   .example(
     '$0 test create https://calibreapp.com --location=Sydney --device=iPhone8 --connection=good3G',
     'Run a test on calibreapp.com from Sydney emulating an iPhone 8 with a 3G connection'
   )
-  .fail((_msg, err, yargs) => {
+  .fail((message, error, yargs) => {
     console.error(
-      chalk.bold.red(err),
+      '\n\n',
+      chalk.bold.red(message ? message : error),
       '\n\n',
       chalk.bold('--- Command documentation\n'),
-      yargs.help(),
-      '\n\n',
-      chalk.bold('--- Stack trace below')
+      yargs.help()
     )
 
-    process.exitCode = 1
-    if (err) throw err // Throw the trace, likley a unhandled promise error
+    if (error && process.env.DEBUG) {
+      console.error('\n\n', chalk.bold('--- Stack trace below'))
+      throw error
+    }
+
+    process.exit(1)
   })
   .epilogue(
     `For more information on Calibre, see ${process.env.CALIBRE_HOST}.`