+++ src/cli/site/get-pulse-metrics.js
@@ -4,30 +4,34 @@ const json2csv = require('json2csv')
 const { humaniseError } = require('../../utils/api-error')
 const { pulse } = require('../../api/snapshot-metrics')
 const formatPulseTimeline = require('../../views/pulse-timeline')
+const { options } = require('../../utils/cli')
 
 const formatCSV = payload => {
   let data = []
 
-  payload.page.timeseries.series.forEach(series => {
+  const timeseries = payload.page ? payload.page.timeseries : payload.timeseries
+
+  timeseries.series.forEach(series => {
     series.sets.forEach(set => {
       const profile = payload.testProfiles.find(
-        profile => profile.id == set.profile.id
+        profile => profile.uuid == set.profile.uuid
       )
 
       set.values.forEach(value => {
-        const snapshot = payload.page.timeseries.snapshots.find(
+        const snapshot = timeseries.snapshots.find(
           snapshot => snapshot.id === value.snapshot
         )
 
         data.push({
           Timestamp: snapshot.createdAt,
-          PageName: payload.page.name,
-          PageURL: payload.page.url,
+          PageUuid: set.page.uuid,
+          PageName: set.page.name,
+          PageURL: set.page.url,
           MetricName: series.metric.name,
           MetricLabel: series.metric.label,
           MetricValue: value.value,
           SnapshotSequenceId: snapshot.sequenceId,
-          TestProfileId: set.profile.id,
+          TestProfileUuid: set.profile.uuid,
           TestProfileName: set.profile.name,
           DeviceName: profile.device ? profile.device.title : null,
           BandwidthName: profile.bandwidth ? profile.bandwidth.title : null,
@@ -41,14 +45,14 @@ const formatCSV = payload => {
 
   const fields = [
     'Timestamp',
+    'PageUuid',
     'PageName',
     'PageURL',
     'MetricName',
     'MetricLabel',
     'MetricValue',
-    'SnapshotId',
     'SnapshotSequenceId',
-    'TestProfileId',
+    'TestProfileUuid',
     'TestProfileName',
     'DeviceName',
     'BandwidthName',
@@ -86,7 +90,7 @@ const main = async args => {
     console.log(formatPulseTimeline(tests))
   } catch (e) {
     if (args.json) return console.error(e)
-    if (args.csv) return console.error(`Error\n${e}`)
+    if (args.csv) return console.error('Error', e)
 
     spinner.fail(humaniseError(e))
     throw new Error(humaniseError(e))
@@ -98,12 +102,8 @@ module.exports = {
   describe: 'Get a recent timeline of metrics for a given page',
   builder: yargs => {
     yargs.options({
-      site: {
-        demandOption: true,
-        describe: 'The identifying slug of a site'
-      },
+      site: options.site,
       page: {
-        demandOption: true,
         describe: 'The identifying uuid of a page'
       },
       metrics: {
@@ -111,8 +111,8 @@ module.exports = {
         describe:
           'A list of metrics to return. eg: --metrics=first-meaningful-paint first-interactive or use multiple --metrics flags for each metric'
       },
-      json: { describe: 'Return pulse data as JSON' },
-      csv: { describe: 'Return pulse data as CSV' },
+      json: options.json,
+      csv: options.csv,
       '30-day': {
         describe:
           'Get the last 30 days of metrics (without this flag, the default is 7 days)'