+++ src/cli/site/test-profiles.js
@@ -4,6 +4,7 @@ const columnify = require('columnify')
 
 const { list } = require('../../api/test-profile')
 const { humaniseError } = require('../../utils/api-error')
+const { options } = require('../../utils/cli')
 
 const main = async args => {
   let index
@@ -53,8 +54,8 @@ module.exports = {
   handler: main,
   builder: yargs => {
     yargs.options({
-      site: { demandOption: true, describe: 'The identifying slug of a site' },
-      json: { describe: 'Return the list of test profiles as JSON' }
+      site: options.site,
+      json: options.json
     })
   }
 }