+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 1.2.0 - 2019-02-19
+
+- Added `calibre test download-artifact <uuid>` command.
+- Added `calibre site download-snapshot-artifacts --site=<slug> <id>` command.
+- Changed `calibre site delete-snapshot` to use `id` rather than `iid`.
+- Changed `calibre site get-pulse-metrics` to default to returning data for all pages. (Previously the --page flag was required)
+- Changed internal queries to use `hasRecentlyCompletedSnapshots` rather than `hasCompletedSnapshots`
+- Added `calibre request --query=<graphql query>` command.
+- Updated node API to use `GraphQL.request`
+- Added mock server and utilities for testing
+
 ### 1.1.1 - 2018-08-22
 
 - Fixed `--csv` output (#26)